Move schema and fix user aggregation
